chore(document/metadata): fix sequence type

+++ document_metadata.go
@@ -10,7 +10,7 @@ type (
 		Permissions []Permissions `nimona:"permissions,omitempty"`
 		Timestamp   uint          `nimona:"timestamp,omitempty"` // TODO: use time.Time
 		Root        *DocumentID   `nimona:"root,omitempty"`
-		Sequence    uint          `nimona:"sequence,omitempty"`
+		Sequence    uint64        `nimona:"sequence,omitempty"`
 		Signature   *Signature    `nimona:"_signature,omitempty"`
 	}
 	Permissions struct {
